Bug Fix - feature/create section - The date validation #22 and refactoring of voting factory.
