+++ backend/src/Features/VotingSection/Domain/ValueObjects/VotingOptionUuid.php
@@ -14,25 +14,25 @@ use JetBrains\PhpStorm\Pure;
 class VotingOptionUuid
 {
     public function __construct(
-        private string $value
+        private ?string $value
     )
     {
     }
 
     /**
-     * @return string
+     * @return string|null
      * @codeCoverageIgnore
      */
-    public function getValue(): string
+    public function getValue(): ?string
     {
         return $this->value;
     }
 
     /**
-     * @param string $value
+     * @param string|null $value
      * @codeCoverageIgnore
      */
-    public function setValue(string $value): void
+    public function setValue(?string $value): void
     {
         $this->value = $value;
     }
